Updated Input, Added Looping Sprite
Input is now handled by passing an enum defined in the source activity
rather than an action pointing to a void returning function in the
activity. This enum is then passed back to the activity when the
associated input(s) are triggered, where it is handled by an overridable
function HandleInput. Looping sprites have been added, and are capable
of loading in a sprite sheet with a set frame per sprite.